requirements.txt
fabric
cuisine
fabtools
fexpect
django
beautifulsoup4
mechanize
south
django-json-field
